Team52-ImperiumBuild/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
branca==0.8.1
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
earthengine-api==1.6.10
fastapi==0.118.0
folium==0.20.0
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-genai==1.41.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
numpy==2.3.3
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
xgboost==3.0.5
xyzservices==2025.4.0
